+++ pubspec.yaml
@@ -30,6 +30,7 @@ dev_dependencies:
     sdk: flutter
   mockito: ^5.1.0
   build_runner: ^2.1.8
+  mobx_codegen: ^2.0.7+2
   flutter_lints: ^2.0.1
